Splitting runtime & simplifying metadata publisher
