+++ lib/mumukit.rb
@@ -34,7 +34,9 @@ require_relative 'mumukit/with_command_line'
 require_relative 'mumukit/version'
 require_relative 'mumukit/content_type'
 require_relative 'mumukit/hook'
-require_relative 'mumukit/runtime'
+require_relative 'mumukit/runtime/info'
+require_relative 'mumukit/runtime/shortcuts'
+require_relative 'mumukit/runtime/runtime'
 require_relative 'mumukit/test_compiler/file_test_compiler'
 require_relative 'mumukit/test_compiler/mashup_test_compiler'
 require_relative 'mumukit/test_runner/isolated_environment'
