+++ lib/mumukit/runtime/runtime.rb
@@ -1,36 +1,10 @@
 class Mumukit::Runtime
+  include Mumukit::RuntimeShortcuts
+  include Mumukit::RuntimeInfo
 
   def initialize(config)
     @config = config
     @hook_classes = {}
-  end
-
-  def feedback_runner
-    new_hook :FeedbackRunner
-  end
-
-  def expectations_runner
-    new_hook :ExpectationsRunner
-  end
-
-  def test_compiler
-    new_hook :TestCompiler
-  end
-
-  def test_runner
-    new_hook :TestRunner
-  end
-
-  def query_runner
-    new_hook :QueryRunner
-  end
-
-  def request_validator
-    new_hook :RequestValidator
-  end
-
-  def metadata_publisher
-    new_hook :MetadataPublisher
   end
 
   def hook_defined?(hook_name)
