+++ lib/mumukit/server/test_server_app.rb
@@ -16,8 +16,8 @@ class Mumukit::TestServerApp < Sinatra::Base
     set :config_filename, 'config/production.yml'
   end
 
-  config = YAML.load_file(settings.config_filename) rescue nil
-  server = Mumukit::TestServer.new(config)
+  runtime_config = YAML.load_file(settings.config_filename) rescue nil
+  server = Mumukit::TestServer.new(runtime_config)
 
   helpers do
     def parse_request
